+++ changelog.md
@@ -14,6 +14,7 @@ TODO:
 
 *	2020.4.24
 
+	+	lilyDocument.ts: term of PostEvent added.
 	+	lilyDocument.ts: term of Tempo added.
 	+	lilyDocument.ts: term of ContextedMusic added.
 	+	lilyDocument.ts: term of SimultaneousList added.